requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
diff-match-patch==20121119
Django==1.9.11
django-crispy-forms==1.6.1
django-extensions==1.7.5
django-filter==1.0.1
django-import-export==0.5.1
django-report-builder==3.4.4
django-simple-history==1.8.2
djangorestframework==3.5.3
enum34==1.1.6
et-xmlfile==1.0.1
html5lib==1.0b8
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
jdcal==1.3
MySQL-python==1.2.5
openpyxl==2.3.5
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pydot==1.2.3
Pygments==2.1.3
pygraphviz==1.3.1
pyparsing==1.5.7
PyPDF2==1.26.0
python-dateutil==2.6.0
reportlab==3.4.0
simplegeneric==0.8.1
six==1.10.0
tablib==0.11.3
traitlets==4.3.1
wcwidth==0.1.7
webencodings==0.5
xhtml2pdf==0.0.6
xlwt==1.2.0
XlsxWriter==0.9.8